Adjusted matchPercentage to work when the lists have a difference number of answers + more unit tests for check matches

+++ app/src/main/java/comp3350/losr/business/AccessMatches.java
@@ -30,7 +30,7 @@ public class AccessMatches {
         List<User> potentialMatches = userAccess.getGenderedUsers();
         User currentUser = dataAccess.getCurrentUser();
 
-        int matchCheck;
+        float matchCheck;
 
         for(int i = 0; i < potentialMatches.size(); i++)
         {
